Includes full arraylike implementation for pwrigid and rigid registration, still need to figure our serialization
