+++ masknmf/motion_correction/strategies.py
@@ -12,6 +12,7 @@ import masknmf
 from masknmf.utils import torch_select_device, display
 from .registration_methods import register_frames_rigid, register_frames_pwrigid
 from masknmf.utils import Serializer
+from masknmf.arrays.array_interfaces import ArrayLike
 
 
 class MotionCorrectionStrategy:
@@ -96,6 +97,7 @@ class MotionCorrectionStrategy:
             reference_movie_frames: np.ndarray,
             target_movie_frames: Optional[np.ndarray] = None,
     ) -> tuple[np.ndarray, np.ndarray]:
+        ## TODO: Decide whether this will really ever be needed
 
         """
         Apply motion correction. Reference frames is the set of frames that we align to the template (to learn
@@ -207,6 +209,11 @@ class MotionCorrectionStrategy:
         ]
         return slice_list
 
+    def motion_correct(self,
+                       reference_movie: ArrayLike,
+                       target_movie: ArrayLike) -> ArrayLike:
+        pass
+
 
 class DummyMotionCorrector(MotionCorrectionStrategy):
     
@@ -232,473 +239,5 @@ class DummyMotionCorrector(MotionCorrectionStrategy):
                          num_iterations: int = 3,
                          ):
         self._template = None
-
-
-class RigidMotionCorrector(MotionCorrectionStrategy, Serializer):
-
-    _serialized = {
-        "max_shifts",
-        "template",
-        "pixel_weighting",
-        "batch_size"
-    }
-    
-    def __init__(
-            self,
-            max_shifts: tuple[int, int] = (15, 15),
-            template: Optional[np.ndarray] = None,
-            pixel_weighting: Optional[np.ndarray] = None,
-            batch_size: int = 200,
-            device: str = "auto",
-    ):
-        super().__init__(template, batch_size=batch_size, device=device)
-
-        self._max_shifts = max_shifts
-        self._pixel_weighting = pixel_weighting
-
-    @property
-    def max_shifts(self) -> tuple[int, int]:
-        """max allowed shift in [row, cols] dim"""
-        return self._max_shifts
-
-    @max_shifts.setter
-    def max_shifts(self, value: tuple[int, int]):
-        value = self._validate_tuple_int_int("max_shifts", value)
-
-        self._template = None
-        self._max_shifts = value
-
-    @property
-    def pixel_weighting(self) -> None | np.ndarray:
-        if self._pixel_weighting is not None:
-            return self._pixel_weighting
-        else:
-            return None
-
-    def _correct_singlebatch(
-            self,
-            reference_frames: np.ndarray,
-            target_frames: Optional[np.ndarray] | None,
-    ) -> tuple[np.ndarray, np.ndarray]:
-
-        if self.template is None:
-            raise ValueError(
-                "Template is uninitialized"
-            )
-
-        #Move appropriate data to cuda
-        if target_frames is not None:
-            target_frames = torch.from_numpy(target_frames).to(self.device).float()
-        reference_frames = torch.from_numpy(reference_frames).to(self.device).float()
-        template = torch.from_numpy(self.template).to(self.device).float()
-        if self.pixel_weighting is not None:
-            pixel_weighting = torch.from_numpy(self.pixel_weighting).to(self.device).float()
-        else:
-            pixel_weighting = None
-
-        outputs = register_frames_rigid(
-            reference_frames,
-            template,
-            self.max_shifts,
-            target_frames=target_frames,
-            pixel_weighting=pixel_weighting
-        )
-        return outputs[0].cpu().numpy(), outputs[1].cpu().numpy()
-
-
-class PiecewiseRigidMotionCorrector(MotionCorrectionStrategy, Serializer):
-    
-    _serialized = {
-        "num_blocks",
-        "overlaps",
-        "max_rigid_shifts",
-        "max_deviation_rigid",
-        "template",
-        "pixel_weighting",
-        "batch_size"
-    }
-    
-    def __init__(
-            self,
-            num_blocks: tuple[int, int] = (12, 12),
-            overlaps: tuple[int, int] = (5, 5),
-            max_rigid_shifts: tuple[int, int] = (15, 15),
-            max_deviation_rigid: tuple[int, int] = (2, 2),
-            template: Optional[np.ndarray] = None,
-            pixel_weighting: Optional[np.ndarray] = None,
-            batch_size: int = 200,
-            device: str = "auto",
-    ):
-        super().__init__(template, batch_size=batch_size, device=device)
-        self._num_blocks = num_blocks
-        self._overlaps = overlaps
-        self._max_rigid_shifts = max_rigid_shifts
-        self._max_deviation_rigid = max_deviation_rigid
-        self._pixel_weighting = pixel_weighting.astype('float') if pixel_weighting is not None else None
-
-    @property
-    def num_blocks(self) -> tuple[int, int]:
-        """
-        Number of blocks that the image plane is split into, [rows, cols].
-        Motion is estimated in each block and then interpolated in 2D space across the entire image plane.
-        """
-        return self._num_blocks
-
-    @num_blocks.setter
-    def num_blocks(self, value):
-        value = self._validate_tuple_int_int("num_blocks", value)
-
-        self._template = None
-        self._num_blocks = value
-
-    @property
-    def pixel_weighting(self) -> None | np.ndarray:
-        return self._pixel_weighting
-
-    @property
-    def overlaps(self) -> tuple[int, int]:
-        """Number of pixels that overlap between adjacent blocks"""
-        return self._overlaps
-
-    @overlaps.setter
-    def overlaps(self, value):
-        value = self._validate_tuple_int_int("overlaps", value)
-
-        self._template = None
-        self._overlaps = value
-
-    @property
-    def max_rigid_shifts(self) -> tuple[int, int]:
-        """maximum shift for rigid iterations, [rows, cols] before piece-wise rigid registration"""
-        return self._max_rigid_shifts
-
-    @max_rigid_shifts.setter
-    def max_rigid_shifts(self, value: tuple[int, int]):
-        value = self._validate_tuple_int_int("max_rigid_shifts", value)
-
-        self._template = None
-        self._max_rigid_shifts = value
-
-    @property
-    def max_deviation_rigid(self) -> tuple[int, int]:
-        """maximum shift allowed in each piece-wise block relative to the rigid registered frame"""
-        return self._max_deviation_rigid
-
-    @max_deviation_rigid.setter
-    def max_deviation_rigid(self, value):
-        value = self._validate_tuple_int_int("max_deviation_rigid", value)
-
-        self._template = None
-        self._max_deviation_rigid = value
-
-    def _correct_singlebatch(
-            self,
-            reference_frames: np.ndarray,
-            target_frames: Optional[np.ndarray],
-    ) -> tuple[np.ndarray, np.ndarray]:
-
-        if self.template is None:
-            raise ValueError(
-                "Template is uninitialized"
-            )
-
-        if target_frames is not None:
-            target_frames = torch.from_numpy(target_frames).to(self.device).float()
-
-        reference_frames = torch.from_numpy(reference_frames).to(self.device).float()
-        
-        template = torch.from_numpy(self.template).to(self.device).float()
-        if self.pixel_weighting is not None:
-            pixel_weighting = torch.from_numpy(self.pixel_weighting).to(self.device).float()
-        else:
-            pixel_weighting = None
-
-        outputs = register_frames_pwrigid(
-            reference_frames.to(self.device),
-            template,
-            self.num_blocks,
-            self.overlaps,
-            self.max_rigid_shifts,
-            self.max_deviation_rigid,
-            target_frames=target_frames,
-            pixel_weighting=pixel_weighting
-        )
-
-        return outputs[0].cpu().numpy(), outputs[1].cpu().numpy()
-
-    def compute_template(
-            self,
-            frames: masknmf.ArrayLike | masknmf.LazyFrameLoader,
-            num_splits_per_iteration: int = 10,
-            num_frames_per_split: int = 200,
-            num_iterations: int = 1,
-    ):
-        rigid_strategy = RigidMotionCorrector(
-            self.max_rigid_shifts,
-            template=self.template.cpu().numpy() if self.template is not None else None,
-            pixel_weighting=self.pixel_weighting,
-            batch_size=self.batch_size,
-            device=self.device,
-        )
-
-        rigid_strategy.compute_template(
-            frames,
-        )
-
-        self._template = rigid_strategy.template
-
-        super().compute_template(
-            frames,
-            num_splits_per_iteration=num_splits_per_iteration,
-            num_frames_per_split=num_frames_per_split,
-            num_iterations=num_iterations,
-        )
-        torch.cuda.empty_cache()
-
-
-class GradientMotionCorrector(MotionCorrectionStrategy, Serializer):
-    """
-    This is a motion corrector designed to correct small (<2 pixel) jitter. Primarily for culture imaging data
-    """
-    _serialized = {
-        "template",
-        "batch_size"
-    }
-    def __init__(
-            self,
-            template: np.ndarray | torch.Tensor | None = None,
-            batch_size: int = 200,
-            pixel_weighting: np.ndarray | torch.Tensor | None = None,
-            device: str = "auto",
-    ):
-        self._device = torch_select_device(device)
-        self.batch_size = batch_size
-        self._pixel_weighting = None
-        self._template = None
-
-        self.pixel_weighting = pixel_weighting
-        self.template = template
-
-
-    @property
-    def dtype(self):
-        """
-        What dtype is used for computations
-        """
-        return torch.float32
-
-    @property
-    def batch_size(self) -> int:
-        """get or set the batch size, the number of frames sent to the GPU in batches for motion correction"""
-        return self._batch_size
-
-    @batch_size.setter
-    def batch_size(self, value: int):
-        if not isinstance(value, int):
-            raise ValueError(f"`batch_size` must be an <int>, you passed: {value}")
-        self._batch_size = value
-
-    @property
-    def template(self) -> None | torch.Tensor:
-        return self._template
-
-    @template.setter
-    def template(self, new_template: None | np.ndarray | torch.Tensor):
-        if new_template is None:
-            self._template = new_template
-        elif isinstance(new_template, (np.ndarray, torch.Tensor)):
-            self._template = torch.as_tensor(new_template, dtype=self.dtype, device=self.device)
-            self._compute_gradient_matrix()
-        else:
-            raise TypeError(f"template should be None, np.ndarray, or torch.Tensor")
-
-    @property
-    def pixel_weighting(self) -> None | torch.Tensor:
-        return self._pixel_weighting
-
-    @pixel_weighting.setter
-    def pixel_weighting(self, weighting: np.ndarray | torch.Tensor):
-        if weighting is None:
-            self._pixel_weighting = None
-        elif isinstance(weighting, (np.ndarray, torch.Tensor)):
-            self._pixel_weighting = torch.as_tensor(weighting, dtype=self.dtype, device=self.device)
-        else:
-            raise TypeError(f"pixel_weighting should be None, np.ndarray, or torch.Tensor")
-
-    def correction_routine(
-            self,
-            reference_frames: np.ndarray,
-            target_frames: Optional[np.ndarray] = None,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """
-        Basic model: we want to register a frame F_1 to a template frame, F_0:
-
-        F_0(x, y) + grad_F0 @ w = alpha * F_1(x, y)
-        where w is a 2D vector of gradient coefficients indicating the size of the shift w.r.t. the template F0.
-        alpha is a scalar (to account for scale differences between F_0 and F_1 because of things like photobleaching)
-
-        This models assumes that the shift between F_0(x, y) and alpha * F_1(x, y) can be explained by the gradient of F0.
-
-        By projecting both sides of this equation to the subspace orthogonal to F_0, we can solve for the vector
-        w / alpha:
-
-        w/alpha = projection of Orth_F0(F1) onto Orth_F0(grad_F0)
-
-        We can then estimate alpha as well as the least squares coefficient relating F_0 and F_1. "w" describes the gradient step
-        in each dimension at the scale of the template image.
-
-        Note that the actual correction applied to frame F_1(x, y) is:
-
-        F_1(x, y) - gradF_0 @ (w / alpha)
-
-        Where this approach breaks down:
-        - Lots of noise relative to the prominent features of the template/reference image
-        - Large dF/F
-
-        Args:
-            reference_frames (np.ndarray): Shape (num_frames, height, width). Frames used to decide gradient shifts in each dimension
-            target_frames (np.ndarray): Shape (num_frames, height, width). Frames to which the correction is applied
-                If None, motion correction is applied to reference_frames
-        Returns:
-            Corrected frame (torch.Tensor): Motion corrected frame. Shape (num_frames, height, width)
-            Shift (torch.Tensor): The shift vectors in gradient space, at the scale of the template image; this is the "w" vector
-            alpha (torch.Tensor): The scale mismatch between the reference frame and template. w / alpha gives the actual gradient coefficients
-                applied to do the correction, as mentioned above.
-        """
-
-        if self.template is None:
-            raise ValueError(
-                "Template is uninitialized"
-            )
-
-        if target_frames is not None:
-            target_frames = torch.as_tensor(target_frames, device=self.device, dtype=self.dtype)
-
-        reference_frames = torch.as_tensor(reference_frames, device=self.device, dtype=self.dtype)
-        num_frames, height, width = reference_frames.shape
-        gradient = torch.as_tensor(self._gradient, device=self.device, dtype=self.dtype)
-        gradient_projector = torch.as_tensor(self._gradient_projector, device=self.device, dtype=self.dtype)
-
-        ref_flat = reference_frames.permute(1, 2, 0).reshape(-1, num_frames)  # [H*W, F]
-        if target_frames is not None:
-            target_flat = target_frames.permute(1, 2, 0).reshape(-1, num_frames)
-        else:
-            target_flat = None
-        gradient_step = gradient_projector @ ref_flat  # [2, F]
-        net_correction = gradient @ gradient_step
-
-        if target_frames is not None:
-            result = target_flat - net_correction
-        else:
-            result = ref_flat - net_correction
-
-        """
-        The actual "gradient step" from the observed data --> template requires more care: need to 
-        make sure the mean image matches the scale of the image in question (photobleaching etc. can affect this). 
-        If they don't match, the gradients will be "off" by some multiplicative factor. 
-
-        So we do a least squares regression of the observed data onto the mean image
-        """
-        template = self.template
-        if self.pixel_weighting is not None:
-            template = template * self.pixel_weighting
-        alpha = torch.sum(reference_frames * template[None, ...], dim=(1, 2)) / torch.sum(reference_frames ** 2,
-                                                                                          dim=(1, 2))
-        shift = gradient_step * alpha[None, :]
-
-        shift = torch.nan_to_num(shift, nan=0.0, posinf=0.0, neginf=0.0).permute(1, 0) #Shape (frames, 2)
-
-        result = result.reshape(height, width, num_frames).permute(2, 0, 1)
-
-        return result, shift, alpha
-
-    @property
-    def gradient(self) -> torch.Tensor | None:
-        """
-        A (pixels, 2) tensor describing the gradients in X and Y dimensions
-        """
-        return self._gradient
-
-    @property
-    def gradient_projector(self) -> torch.Tensor | None:
-        return self._gradient_projector
-    def compute_gradient_step(self,
-                              shifts: torch.Tensor,
-                              alpha: torch.Tensor):
-        """
-        The shifts computed by the gradient strategy are at the scale of the template image. This routine computes the
-        gradient coefficients that are used to correct the i-th frame of data, this routine returns w / alpha
-
-        Args:
-            shifts (torch.Tensor): Shape (num_frames, 2). The gradient shifts in each dimension, at the scale of the template
-            alpha (torch.Tensor): Shape (num_frames). The scale mismatch between the template and the frames
-        """
-        gradient_steps = shifts / alpha[:, None]
-        gradient_steps = torch.nan_to_num(gradient_steps, nan=0.0, posinf=0.0, neginf=0.0)
-        return gradient_steps
-
-    def _correct_singlebatch(
-            self,
-            reference_frames: np.ndarray,
-            target_frames: Optional[np.ndarray],
-            apply_alpha_correction: bool = True
-    ) -> tuple[np.ndarray, np.ndarray]:
-        result, shifts, _ = self.correction_routine(reference_frames,
-                                                         target_frames)
-
-
-        return result.cpu().numpy(), shifts.cpu().numpy()
-
-    def compute_template(
-            self,
-            frames: masknmf.ArrayLike | masknmf.LazyFrameLoader
-    ):
-        self.template = frames[0].squeeze(0)
-
-    def _compute_gradient_matrix(self):
-        """
-        Computes the gradient design matrix used for the motion correction
-        :return:
-        """
-        f0 = self.template
-        dfdx_orig = torch.nn.functional.pad(
-            (f0[:, 2:] - f0[:, :-2]) / 2, (1, 1), mode="constant"
-        )  # x-gradient [H, W]
-        dfdy_orig = torch.nn.functional.pad(
-            (f0[2:, :] - f0[:-2, :]) / 2, (0, 0, 1, 1), mode="constant"
-        )  # y-gradient [H, W]
-
-
-
-        if self.pixel_weighting is not None:
-            f0 = f0 * self.pixel_weighting
-            mask =self.pixel_weighting
-            dfdx = dfdx_orig * mask
-            dfdy = dfdy_orig * mask
-        else:
-            dfdx = dfdx_orig
-            dfdy = dfdy_orig
-
-        f0_flat = f0.flatten()  # [H*W]
-        f0_norm_sq = f0_flat @ f0_flat
-
-        def orthogonalize_against_template(v):
-            """
-            This routine orthogonalizes against the mean image. This is important because it
-            makes the shift estimation insensitive to
-            """
-            return v - (f0_flat @ v) / f0_norm_sq * f0_flat
-
-        dfdx_flat = orthogonalize_against_template(dfdx.flatten())
-        dfdy_flat = orthogonalize_against_template(dfdy.flatten())
-
-        A = torch.stack([dfdx_flat, dfdy_flat], dim=1).float()  # [H*W, 2]
-        AtA_inv = torch.linalg.inv(A.T @ A)  # [2, 2]
-        A_projector = AtA_inv @ A.T  # [2, H*W]
-
-        self._gradient = torch.stack([dfdx_orig.flatten(), dfdy_orig.flatten()], dim=1).to(self.dtype).to(self.device) ## The design matrix doesn't change
-        self._gradient_projector = A_projector.to(self.dtype).to(self.device) ## The projector can be orthogonalized w.r.t. f0
-
-
